chain: resolve races in Rescan and NotifyReceived
Implements a concurrent test on Rescan and NotifyReceived.
Modifies the rescanQuit prop to be a channel of channels with
a buffer of 1.  If a rescan is running and Rescan is called,
the channel in the buffer will be dequeued and closed.  A new
channel will be put on the channel in it's place.

Wraps the handlers in closures that use the new channel.  Avoids
a data race in constantly switching out rescanQuit because
channels are concurrent safe.