chain: use rescanQuitCh
Use a channel of channels to hold the rescanQuit channel that must
be closed to signal a rescan struct to stop.  A channel of channels
provides a concurrent safe way to access the single object instead
of using mutexes to access rescanQuit as a property of the client.

Add rescanQuit as a parameter to each of the notification handlers.